refactor: remove EnqueuedCallSimulator
